+++ src/pages/rolling/paper/components/RollingPaperItem.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import CardList from '@Paper/components/RollingPaperItem.style';
-import Card from '@Components/card/Card';
+import Card from '@Components/messageCard/MessageCard';
 
 function RollingPaperItem({ data, onClickModal, getPaperData }) {
   const handleOpenModal = () => {
